:recycle: 공통 Title 컴포넌트 분리
